requirements.txt
pyotp==2.9.0
qrcode[pil]==7.4.2
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
cryptography==41.0.7
pywin32==306
pystray==0.19.5
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
packaging==23.2
pyinstaller==6.3.0